camera 탭에 사진 찍는 기능과 갤러리 기능을 같이 넣음 갤러리는 horizontal 하게 scroll할 수 있도록 함 하지만 fragment를 불러오는 과정에서 상단과 하단에 여백이 생기는 것으로 보임
