+++ .gradle/config.properties
@@ -1,2 +1,2 @@
-#Sun Jun 30 23:07:19 KST 2024
-java.home=/Applications/Android Studios.app/Contents/jbr/Contents/Home
+#Mon Jul 01 20:31:54 KST 2024
+java.home=C\:\\Program Files\\Android\\Android Studio1\\jbr
