+++ local.properties
@@ -5,5 +5,5 @@
 # For customization when using a Version Control System, please read the
 # header note.
 #Fri Jun 28 13:27:16 KST 2024
-#sdk.dir=C\:\\Users\\82109\\AppData\\Local\\Android\\Sdk
-sdk.dir=/Users/kim-eunseo/Library/Android/sdk
+sdk.dir=C\:\\Users\\82109\\AppData\\Local\\Android\\Sdk
+#sdk.dir=/Users/kim-eunseo/Library/Android/sdk
